third_party/blink/web_tests/editing/pasteboard/drag-and-drop-objectimage-contenteditable-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
This test verifies dropping an image does nothing. To test manually, drag the image and drop it in the rect with a blue border.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS imageIsInserted is false
PASS successfullyParsed is true

TEST COMPLETE
